Revert "merge main"
This reverts commit 5da5e282d53f7f717ab4c0b9918053b51f2f05c5.

+++ src/Api/Billing/Models/Responses/PaymentMethodResponse.cs
@@ -4,7 +4,7 @@ using Bit.Core.Billing.Tax.Models;
 namespace Bit.Api.Billing.Models.Responses;
 
 public record PaymentMethodResponse(
-    decimal AccountCredit,
+    long AccountCredit,
     PaymentSource PaymentSource,
     string SubscriptionStatus,
     TaxInformation TaxInformation)
